[Extensibility][SDK] Correct `print_kv_cache` functions. (#53460)
Sets `num_dps_outputs=0` as properly expected and corrects the test to match the proper dtype.

Testing:
In [Internal link] we can run `//SDK/integration-test/API:mojo/test_model.mojo.test` and get it to pass.
MAX_INTEGRATION_TESTS_REV_ID: 8a39d798d468b3c1d0c7dad224f370c36b19d778

+++ max/tests/integration/pipelines/python/nn/kv_cache/test_print_kv_cache.py
@@ -79,7 +79,7 @@ def test_print_kv_cache() -> None:
         "print_kv_cache",
         forward=PrintKVCacheModel(fetch_layer, kv_params, layer_idx=0),
         input_types=[
-            TensorType(dtype=DType.int64, shape=[batch_size]),
+            TensorType(dtype=DType.uint32, shape=[batch_size]),
             *kv_manager.input_symbols()[0],
         ],
     )
